Allow manual cycle deposits into user account (#6)
Two new functions are added to the public interface of the Tip Jar canister:

   depositCyclesFor: (principal) -> (Result);
   findUserByAccount: (text) -> (opt principal) query;

depositCyclesFor will deposit all cycles sent along this call to the given user's account.

findUserByAccount will return the user's principal given its account id used for ICP deposit. The caveat is that this may run out of execution limit. It will become a problem in the future, but for now it should be just fine.